Add files via upload
Binary Search is an efficient algorithm for finding the position of a target value within a sorted array. It works by repeatedly dividing the search interval in half. If the value of the target is less than the middle element, the search continues in the lower half; if it’s greater, it continues in the upper half. This process continues until the target value is found or the interval is reduced to zero.

Binary Search is much faster than linear search for sorted arrays, with a time complexity of O(log n), where n is the number of elements in the array.