Add files via upload
Linear Search is a simple search algorithm that looks for a specific element in a list or array by checking each element sequentially, one by one. It is one of the most basic and straightforward searching algorithms, making it easy to understand and implement.

Linear search works in both unsorted and sorted lists, as it does not rely on the arrangement of elements. The algorithm starts at the beginning of the list and continues until the desired element is found or the end of the list is reached.