Add files via upload
A stack is a linear data structure that operates on a Last In, First Out (LIFO) principle. This means that the last element added to the stack is the first one to be removed. It is analogous to a stack of plates: the last plate placed on the stack is the first one to be taken off.